requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
chess==1.11.2
dj-database-url==2.3.0
Django==5.2.1
idna==3.10
psycopg2-binary==2.9.10
python-chess==1.999
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
